LIBITD-1470. Updated Dockerfile to run "scripts/cleanup.sh" script
Added dependencies to run the "scripts/cleanup.sh" script against
the "data.csv" file.

The "data.csv" file corresponds to the "scores_20140922revised.csv"
on the "SolrDB Project: SCPA Scores Collection" Confluence page:

https://confluence.umd.edu/display/LIB/SolrDB+Project%3A+SCPA+Scores+Collection

Note: There are data discrepancies between this file and the data on
solr.lib.umd.edu.

Modified "scripts/cleanup.sh":

a) Changed "dos2unix" command to always covert the file (even if it
thinks it is binary), using the "-f" flag, and modified to produce an
output file with the same name as the input file with a ".unix" suffix.
These modifications were made, as they seemed to be an expectation of
the script.

b) Changed the "cvsjoin" command to use the "--no-inference" flag, as
otherwise the zero-padded id files ("0000001") would end up without the
zero-padding ("1") in the output file. Using the flag preserves the
zero-padding.

https://issues.umd.edu/browse/LIBITD-1470